Rename query params files
